Decomp wave 4 (held TU landed): CgsDebugComponentMessageFilter + gxMessageFilterFlags u64 fix
Resolves the ODR blocker that held CgsDebugComponentMessageFilter back in wave 4.
The verify pass was PASS/high; its only issue was a cross-header type conflict on
the CgsDev::Message::gxMessageFilterFlags global: committed CgsLog.h/.cpp declared
it u32 while this TU's CgsMessage.h (DWARF CgsMessage.h:40 typedef uint64_t, and
the X360 asm loads/stores the full 64-bit value) needs u64.

- CgsLog.h / CgsLog.cpp: widen gxMessageFilterFlags u32 -> u64 to match the X360
  ground truth and CgsMessage.h. All readers gate on `(flags & bit)` (width-
  agnostic); compile-checked a representative sample of the ~28 readers
  (CgsVariableManager, BrnGameActions, BrnRaceMode, BrnCarSelectManager,
  BrnGuiHudMessageDirector, CgsRemapDevice, BrnTriggerQueryManager) -- all pass.

- CgsDebugComponentMessageFilter (7 fns): the message-filter debug component
  (GetName/SetFilterName/FilterEnable/Type change callbacks). Edits the 64-bit
  gxMessageFilterFlags; the reinterpret_cast<u32*> low/high-half access models the
  X360's two-u32-register access of the 64-bit value. CgsMessage.h carries the
  FilterFlag (u64) typedef + the KX_FILTER_* category bits.

+++ src/GameShared/GameClasses/Development/Log/CgsLog.cpp
@@ -52,6 +52,6 @@ namespace Log
 
 namespace Message
 {
-    u32 gxMessageFilterFlags = 0xFFFFFFFF;
+    u64 gxMessageFilterFlags = 0xFFFFFFFF;
 }
 }
