+++ src/GameShared/GameClasses/Development/Log/CgsLog.h
@@ -30,6 +30,9 @@ namespace Message
 {
     // Log-category filter. Call sites gate logging on `(gxMessageFilterFlags & bit)`; set broad
     // so the engine's logging is captured. (The real per-category default differs; tune later.)
-    extern u32 gxMessageFilterFlags;
+    // Width is 64-bit: the X360 asm loads/stores the full 64-bit value and the DecFIGS DWARF
+    // (CgsMessage.h:40) types it `uint64_t`. Matches CgsMessage.h's `extern FilterFlag (u64)`
+    // declaration of the same CgsDev::Message::gxMessageFilterFlags symbol (no ODR conflict).
+    extern u64 gxMessageFilterFlags;
 }
 }
